Add useHintPosition hook for dynamic hint positioning logic
