+++ src/components/hint/index.tsx
@@ -8,12 +8,13 @@ import Text from '../text';
 import Image from '../image';
 import Modal from '../modal';
 import TouchableOpacity from '../touchableOpacity';
-import {HintPositions, HintPositionStyle, HintProps, Paddings, Position, TARGET_POSITIONS} from './types';
+import {HintPositions, HintProps, TARGET_POSITIONS} from './types';
 
 import {useDidUpdate} from 'hooks';
 import useHintAnimation from './hooks/useHintAnimation';
 import useHintLayout from './hooks/useHintLayout';
 import useHintAccessibility from './hooks/useHintAccessibility';
+import useHintPosition from './hooks/useHintPosition';
 
 const sideTip = require('./assets/hintTipSide.png');
 const middleTip = require('./assets/hintTipMiddle.png');
@@ -52,6 +53,7 @@ const NewHint = (props: HintProps) => {
   } = props;
 
   const hintRef = useRef<RNView>(null);
+  const isUsingModal = onBackgroundPress && useModal;
 
   const {hintUnmounted, visibleAnimated, animateHint} = useHintAnimation(visible);
   const {targetLayoutState, targetLayoutInWindowState, hintMessageWidth, targetRef, onTargetLayout, setHintLayout} =
@@ -73,8 +75,8 @@ const NewHint = (props: HintProps) => {
       return targetFrame;
     }
 
-    return onBackgroundPress && useModal ? targetLayoutInWindowState : targetLayoutState;
-  }, [onBackgroundPress, useModal, targetLayoutState, targetLayoutInWindowState, targetFrame]);
+    return isUsingModal ? targetLayoutInWindowState : targetLayoutState;
+  }, [isUsingModal, targetLayoutState, targetLayoutInWindowState, targetFrame]);
 
   const setTargetRef = useCallback((ref: ElementRef<typeof RNView>) => {
     targetRef.current = ref;
@@ -88,7 +90,6 @@ const NewHint = (props: HintProps) => {
   }, []);
 
   const showHint = !!targetLayout;
-  const isUsingModal = onBackgroundPress && useModal;
 
   const edgeMargins = useMemo(() => {
     if (edgeMarginsProp !== undefined) {
@@ -97,110 +98,15 @@ const NewHint = (props: HintProps) => {
     return isUsingModal ? DEFAULT_EDGE_MARGINS : 0;
   }, [isUsingModal, edgeMarginsProp]);
 
-  const targetPositionOnScreen = useMemo(() => {
-    if (targetLayout?.x !== undefined && targetLayout?.width) {
-      const targetMidPosition = targetLayout.x + targetLayout.width / 2;
-
-      if (targetMidPosition > containerWidth * (4 / 5)) {
-        return TARGET_POSITIONS.RIGHT;
-      } else if (targetMidPosition < containerWidth * (1 / 5)) {
-        return TARGET_POSITIONS.LEFT;
-      }
-    }
-    return TARGET_POSITIONS.CENTER;
-  }, [targetLayout, containerWidth]);
-
-  const shouldUseSideTip = useMemo(() => {
-    if (!_.isUndefined(useSideTip)) {
-      return useSideTip;
-    }
-
-    return targetPositionOnScreen !== TARGET_POSITIONS.CENTER;
-  }, [useSideTip, targetPositionOnScreen]);
-
-  const tipSize = useMemo(() => {
-    return shouldUseSideTip ? {width: 14, height: 7} : {width: 20, height: 7};
-  }, [shouldUseSideTip]);
-
-  const hintPadding = useMemo(() => {
-    const paddings: Paddings = {paddingVertical: offset, paddingHorizontal: edgeMargins};
-
-    if (shouldUseSideTip && targetLayout?.x !== undefined) {
-      if (targetPositionOnScreen === TARGET_POSITIONS.LEFT) {
-        paddings.paddingLeft = targetLayout.x;
-      } else if (targetPositionOnScreen === TARGET_POSITIONS.RIGHT && targetLayout?.width) {
-        paddings.paddingRight = containerWidth - targetLayout.x - targetLayout.width;
-      }
-    }
-
-    return paddings;
-  }, [targetLayout, containerWidth, shouldUseSideTip, targetPositionOnScreen, offset, edgeMargins]);
-
-  const hintPosition = useMemo(() => {
-    const hintPositionStyle: HintPositionStyle = {alignItems: 'center'};
-
-    if (targetLayout?.x !== undefined) {
-      hintPositionStyle.left = -targetLayout.x;
-    }
-
-    if (position === HintPositions.TOP) {
-      hintPositionStyle.bottom = 0;
-    } else if (targetLayout?.height) {
-      hintPositionStyle.top = targetLayout.height;
-    }
-
-    if (targetPositionOnScreen === TARGET_POSITIONS.RIGHT) {
-      hintPositionStyle.alignItems = Constants.isRTL ? 'flex-start' : 'flex-end';
-    } else if (targetPositionOnScreen === TARGET_POSITIONS.LEFT) {
-      hintPositionStyle.alignItems = Constants.isRTL ? 'flex-end' : 'flex-start';
-    }
-
-    return hintPositionStyle;
-  }, [position, targetLayout, targetPositionOnScreen]);
-
-  const containerPosition = useMemo(() => {
-    if (targetLayout) {
-      return {top: targetLayout.y, left: targetLayout.x};
-    }
-  }, [targetLayout]);
-
-  const tipPosition = useMemo(() => {
-    const tipPositionStyle: Position = {};
-
-    if (position === HintPositions.TOP) {
-      tipPositionStyle.bottom = offset - tipSize.height;
-      !shouldUseSideTip ? (tipPositionStyle.bottom += 1) : undefined;
-    } else {
-      tipPositionStyle.top = offset - tipSize.height;
-    }
-
-    const layoutWidth = targetLayout?.width || 0;
-
-    if (targetLayout?.x !== undefined) {
-      const targetMidWidth = layoutWidth / 2;
-      const tipMidWidth = tipSize.width / 2;
-
-      const leftPosition = shouldUseSideTip ? targetLayout.x : targetLayout.x + targetMidWidth - tipMidWidth;
-      const rightPosition = shouldUseSideTip
-        ? containerWidth - targetLayout.x - layoutWidth
-        : containerWidth - targetLayout.x - targetMidWidth - tipMidWidth;
-
-      switch (targetPositionOnScreen) {
-        case TARGET_POSITIONS.LEFT:
-          tipPositionStyle.left = Constants.isRTL ? rightPosition : leftPosition;
-          break;
-        case TARGET_POSITIONS.RIGHT:
-          tipPositionStyle.right = Constants.isRTL ? leftPosition : rightPosition;
-          break;
-        case TARGET_POSITIONS.CENTER:
-        default:
-          tipPositionStyle.left = targetLayout.x + targetMidWidth - tipMidWidth;
-          break;
-      }
-    }
-
-    return tipPositionStyle;
-  }, [targetLayout, containerWidth, position, targetPositionOnScreen, offset, shouldUseSideTip, tipSize]);
+  const {shouldUseSideTip, targetPositionOnScreen, hintPosition, containerPosition, tipPosition, hintPadding} =
+    useHintPosition({
+      useSideTip,
+      position,
+      offset,
+      targetLayout,
+      containerWidth,
+      edgeMargins
+    });
 
   const hintOffsetForShortMessage = useMemo(() => {
     let hintMessageOffset = 0;
@@ -235,7 +141,7 @@ const NewHint = (props: HintProps) => {
         }
       ]
     };
-  }, [position]);
+  }, [position, visibleAnimated]);
 
   const renderOverlay = () => {
     if (targetLayoutInWindowState && containerPosition?.top && containerPosition?.left) {
@@ -471,6 +377,8 @@ NewHint.defaultProps = {
   useModal: true
 };
 NewHint.positions = HintPositions;
+
+export {HintProps};
 
 // @ts-expect-error
 export default asBaseComponent<HintProps, typeof NewHint>(NewHint);
